merkle-test-fixtures moved
