+++ platform-sdk/platform-apps/tests/PlatformTestingTool/src/main/java/com/swirlds/demo/merkle/map/internal/ExpectedMapUtils.java
@@ -24,10 +24,10 @@ import com.swirlds.demo.platform.PlatformTestingToolState;
 import com.swirlds.demo.platform.SuperConfig;
 import com.swirlds.demo.platform.Triple;
 import com.swirlds.demo.platform.UnsafeMutablePTTStateAccessor;
-import com.swirlds.merkle.map.test.lifecycle.ExpectedValue;
-import com.swirlds.merkle.map.test.lifecycle.LifecycleStatus;
-import com.swirlds.merkle.map.test.lifecycle.TransactionState;
-import com.swirlds.merkle.map.test.pta.MapKey;
+import com.swirlds.merkle.test.fixtures.map.lifecycle.ExpectedValue;
+import com.swirlds.merkle.test.fixtures.map.lifecycle.LifecycleStatus;
+import com.swirlds.merkle.test.fixtures.map.lifecycle.TransactionState;
+import com.swirlds.merkle.test.fixtures.map.pta.MapKey;
 import com.swirlds.platform.listeners.ReconnectCompleteNotification;
 import com.swirlds.platform.system.Platform;
 import java.time.Instant;
